Move Settings usage to lifedrain.py
